Backend/datakom_prosjekt/requirements.txt
asgiref==3.7.2
astroid==3.0.1
azure-core==1.29.5
azure-storage-blob==12.18.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
dill==0.3.7
Django==4.2.6
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
isodate==0.6.1
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pylint==3.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
urllib3==1.26.18
django-cors-headers==4.3.0
